requirements.txt
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
filelock==3.25.0
Flask==2.3.3
fsspec==2026.2.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.6.1
numpy==1.24.3
opencv-python-headless==4.8.1.78
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079066/work
Pillow==10.0.1
requests==2.32.5
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==2.3.7
